Move Grafana state tests to pytest
